feat: add title and meta tags in generate page
